# user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: BookMania User Microservice - OpenAPI 3.0
  description: |-
    We want to develop a social media platform dedicated to book enthusiasts and readers of all kinds, called BookMania. It serves as a virtual library where users can catalog and review the books they've read, as well as discover new titles based on their interests and reading history. The platform allows members to create personalized bookshelves, join discussion groups, and participate in reading challenges. It also provides a space for users to connect with friends, see their recommendations, and follow authors.

  version: 0.0.1
servers:
  - url: https://localhost:8080
    description: Local development server
...
paths:
  /book/{username}:
    post:
      tags:
        - book
      summary: Add a new book to the store
      description: Add a new book to the store
      operationId: addBook
      parameters:
        - in: path
          required: true
          name: username
          description: The user's username
          schema:
            type:
              string
      requestBody:
        description: Create a new book in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: User not found
        '403':
          description: User is not an admin
        '406':
          description: Book cannot be null
        '400':
          description: Book fields are not valid
  /book/search:
    get:
      tags:
        - book
      summary: Search books by author, genre, title, description, series and sort by read_count, alphabetical
      description: Retrieve books based on specified criteria
      parameters:
        - name: author
          in: query
          description: The author of the books to search for
          required: false
          schema:
            type: string
        - name: genre
          in: query
          description: The genre of the books to search for
          required: false
          schema:
            type: string
        - name: title
          in: query
          description: The title of the books to search for
          required: false
          schema:
            type: string
        - name: description
          in: query
          description: The description of the books to search for
          required: false
          schema:
            type: string
        - name: series
          in: query
          description: The series of the books to search for
          required: false
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort the books by a specific field (e.g., read_count, alphabetical)
          required: false
          schema:
            type: string
            enum: [read_count, alphabetical, read_count_reversed, alphabetical_reversed]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
  /book/{id}:
    get:
      tags:
        - book
      summary: Get a book
      description: Get a book by id
      operationId: getBook
      parameters:
        - name: id
          in: path
          description: 'The id of the book that needs getting'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid ID supplied
        '404':
          description: Book not found
  /book/{id}/{username}:
    put:
      tags:
        - book
      summary: Update an existing book
      description: Update an existing book by Id
      operationId: updateBook
      parameters:
        - name: id
          in: path
          description: 'The id of the book that needs changing'
          required: true
          schema:
            type: string
        - in: path
          required: true
          name: username
          description: The user's username
          schema:
            type:
              string
      requestBody:
        description: Update an existent book in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid fields of the updated book
        '403':
          description: Unauthorised user
        '404':
          description: This book does not exist
        '406':
          description: Updated book can not be null
    delete:
      tags:
        - book
      summary: Remove a book from the store
      description: Remove a book from the store by id
      operationId: removeBook
      parameters:
        - name: id
          in: path
          description: 'The id of the book that needs deletion'
          required: true
          schema:
            type: string
        - in: path
          required: true
          name: username
          description: The user's username
          schema:
            type:
              string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '403':
          description: Unauthorised user
        '404':
          description: Book not found
  /book/read/{id}:
    put:
      tags:
        - book
      summary: Read a book
      description: 'Adds 1 to the read count for the given book'
      operationId: readBook
      parameters:
        - name: id
          in: path
          description: 'The id of the book that the user has read'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book successfully read
        '404':
          description: Book not found
        '400':
          description: Book must have an id
  /book:
    post:
      tags:
        - book
      summary: Get Books by IDs
      description: Retrieve books based on provided IDs
      requestBody:
        description: List of book IDs
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
...
components:
  schemas:
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        title:
          type: string
          example: 'The Best Book'
        description:
          type: string
          example: 'The greatest book ever'
        author:
          type: string
          example: 'Bob Bobson'
        reads:
          type: integer
          format: int64
          example: 0
        series:
          type: string
          example: 'The Big Trilogy'
        genres:
          type: array
          items:
            type: string
          example:
            - 'Horror'
            - 'Sci-fi'
            - 'Thriller'
...